Update text render loop to use `StringSlice` array
This removes the need to check for newlines in the text render loop, simplifying it. This will also simplify the application. Rather than requiring application to concatenate all text into a monster string, it can simply give the renderer chunks of text to render, and it has complete control over the positioning of that text.