Introduce `LinkedMemoryStack`
The existing `MemoryStack` allocated more memory by allocating new space, copying the old data to it, then freeing the old spaces. This prohibits anyone from creating pointers to memory allocated from this stack, which is inconvenient. An index could be used instead of a pointer, and probably wouldn't be too bad of a workaround, but again, it's inconvenient.

The `LinkedMemoryStack` works by allocating a brand new stack when the existing stack is too small, and maintaining a link to the old stack. This preserves the validity of pointers and eliminates the need for a copy when allocating more space. At the end of the frame, all stacks are freed, and a new stack is allocated with a size equal to the combined sizes of the old stacks. Over time, the stack grows to a size fitting the application's needs.

This approach ensures that the application does not allocate far more memory than it would ever require at startup, improving startup time and ensuring that it cooperates fairly with other applications on the user's machine.