Remove some overloads for calculating `TextLocation`
These overloads accepted a raw character pointer, rather than a full text location. This means that, for example, some code could save a character pointer, advance the parser forward a few lines, then report an error using this character. This means that the reported line and character number would be wrong. Callers are now required to suppply a full text location, which makes it hard to record an invalid text position.